Bruce/data model bucket (#7988)
* wip

* remove all old models imports by the new librabry ( festival )

* Updating imports lib by => '@blockframes/model'

* correcting error test

* correcting error test 2

* remove file

* correcting review

* correcting error test

* correcting test error

* updated review

* cleaning

* cleaning

* cleaning

* cleaning
