+++ apps/backend-functions/src/contracts.ts
@@ -1,11 +1,10 @@
 import { db } from './internals/firebase';
-import { Contract, Sale } from '@blockframes/contract/contract/+state/contract.model';
 import { Change } from 'firebase-functions';
 import { createNotification, triggerNotifications } from './notification';
 import { createDocumentMeta, getDocument, Timestamp } from './data/internals';
 import { Negotiation } from '@blockframes/contract/negotiation/+state/negotiation.firestore';
 import { getReviewer } from '@blockframes/contract/negotiation/utils';
-import { Organization, NotificationDocument } from '@blockframes/model';
+import { Organization, NotificationDocument, Sale, Contract } from '@blockframes/model';
 
 interface ContractNotificationType {
   sender: 'myOrgAcceptedAContract' | 'myOrgDeclinedAContract', //org who accepted/declined a contract
