+++ apps/backend-functions/src/negotiation.ts
@@ -3,12 +3,11 @@ import { Change, EventContext } from 'firebase-functions';
 import { Negotiation } from '@blockframes/contract/negotiation/+state/negotiation.firestore';
 import { createDocumentMeta, getDocument, Timestamp } from './data/internals';
 import { centralOrgId } from 'env/env.blockframes-ci';
-import { Organization } from '@blockframes/model';
+import { Contract, ContractStatus, Movie, Organization, Sale } from '@blockframes/model';
 import { createNotification, triggerNotifications } from './notification';
 import { getReviewer, isInitial } from '@blockframes/contract/negotiation/utils'
 import { formatDocumentMetaFromFirestore } from "@blockframes/utils/models-meta";
 import { Offer } from '@blockframes/contract/offer/+state';
-import { Contract, ContractStatus, Sale } from '@blockframes/contract/contract/+state/contract.model';
 import { NotificationTypes } from '@blockframes/model';
 
 // KEEP THE OFFER STATUS IN SYNC WITH IT'S CONTRACTS AND NEGOTIATIONS
