+++ apps/backend-functions/src/notification.ts
@@ -13,6 +13,7 @@ import {
   NotificationDocument,
   InvitationDocument,
   PublicInvitation,
+  ContractDocument,
   Screening
 } from '@blockframes/model';
 import { sendMailFromTemplate } from './internals/email';
@@ -54,7 +55,6 @@ import * as admin from 'firebase-admin';
 import { logger } from 'firebase-functions';
 import { NegotiationDocument } from '@blockframes/contract/negotiation/+state/negotiation.firestore';
 import { Offer } from '@blockframes/contract/offer/+state';
-import { ContractDocument } from '@blockframes/contract/contract/+state';
 import { appUrl, supportEmails } from './environments/environment';
 import { getReviewer } from '@blockframes/contract/negotiation/utils';
 // #7946 this may be reactivated later
