+++ apps/catalog/catalog/src/app/marketplace/title/selection/selection.component.ts
@@ -1,7 +1,7 @@
 import { Component, ChangeDetectionStrategy, Optional } from '@angular/core';
 import { Intercom } from 'ng-intercom';
 import { BucketService } from '@blockframes/contract/bucket/+state';
-import { Movie, Bucket } from '@blockframes/model';
+import { Movie, Bucket, Holdback } from '@blockframes/model';
 import { DynamicTitleService } from '@blockframes/utils/dynamic-title/dynamic-title.service';
 import { movieCurrencies } from '@blockframes/utils/static-model';
 import { Observable, Subject, merge } from 'rxjs';
@@ -11,7 +11,6 @@ import { MatDialog } from '@angular/material/dialog';
 import { SpecificTermsComponent } from './specific-terms/specific-terms.component';
 import { OrganizationService } from '@blockframes/organization/+state';
 import { FormControl } from '@angular/forms';
-import { Holdback } from '@blockframes/contract/contract/+state';
 
 @Component({
   selector: 'catalog-selection',
