+++ apps/crm/crm/src/app/crm/pages/offers/shell/contract-form/contract-form.component.ts
@@ -3,9 +3,9 @@ import { ActivatedRoute } from '@angular/router';
 
 // Services
 import { MovieService } from '@blockframes/movie/+state/movie.service';
-import { Movie } from '@blockframes/model';
+import { Contract, Movie } from '@blockframes/model';
 import { Income, IncomeService } from '@blockframes/contract/income/+state';
-import { Contract, ContractService } from '@blockframes/contract/contract/+state';
+import { ContractService } from '@blockframes/contract/contract/+state';
 import { TermService } from '@blockframes/contract/term/+state';
 import { OfferService } from '@blockframes/contract/offer/+state';
