+++ apps/crm/crm/src/app/crm/pages/offers/shell/contract-view/contract-view.component.ts
@@ -10,12 +10,13 @@ import { combineLatest, Subscription } from 'rxjs';
 
 import { IncomeService } from '@blockframes/contract/income/+state';
 import { Term } from '@blockframes/contract/term/+state';
-import { ContractService, Holdback } from '@blockframes/contract/contract/+state';
+import { ContractService } from '@blockframes/contract/contract/+state';
 import { ConfirmInputComponent } from '@blockframes/ui/confirm-input/confirm-input.component';
 import { OfferShellComponent } from '../shell.component';
 import { NegotiationService } from '@blockframes/contract/negotiation/+state/negotiation.service';
 import { Negotiation } from '@blockframes/contract/negotiation/+state/negotiation.firestore';
 import { isInitial } from '@blockframes/contract/negotiation/utils';
+import { Holdback } from '@blockframes/model';
 
 
 @Component({
