+++ libs/contract/src/lib/avails/avails.ts
@@ -8,8 +8,7 @@ import {
   TerritoryValue
 } from '@blockframes/utils/static-model';
 import { BucketTerm, Term } from '../term/+state';
-import { Holdback, Mandate, Sale } from '../contract/+state';
-import { Bucket, BucketContract } from '@blockframes/model';
+import { Holdback, Mandate, Sale, Bucket, BucketContract } from '@blockframes/model';
 import { allOf, exclusivityAllOf, exclusivitySomeOf, someOf } from './sets';
 
 export interface BaseAvailsFilter {
