+++ libs/contract/src/lib/bucket/+state/bucket.service.ts
@@ -11,9 +11,9 @@ import { createDocumentMeta } from '@blockframes/utils/models-meta';
 import { OrganizationService } from '@blockframes/organization/+state';
 import { TermService } from '../../term/+state';
 import { OfferService } from '../../offer/+state';
-import { Bucket, createBucket, createBucketTerm, createBucketContract } from '@blockframes/model';
-import { ContractService, convertDuration } from '../../contract/+state';
+import { ContractService } from '../../contract/+state';
 import { ActiveState, EntityState } from '@datorama/akita';
+import { convertDuration, Bucket, createBucket, createBucketTerm, createBucketContract } from '@blockframes/model';
 
 interface BucketState extends EntityState<Bucket>, ActiveState<string> { }
