+++ libs/contract/src/lib/contract/+state/index.ts
@@ -1,2 +1,1 @@
-export * from './contract.model';
 export * from './contract.service';
