+++ libs/contract/src/lib/contract/+state/utils.ts
@@ -1,5 +1,5 @@
+import { Contract } from "@blockframes/model";
 import { centralOrgId } from "@env";
-import { Contract } from "./contract.firestore";
 
 export function getSeller(contract: Contract) {
   const sellerId = contract.stakeholders.find(stakeholder => {
