+++ libs/contract/src/lib/contract/holdback/form/form.component.ts
@@ -9,11 +9,10 @@ import {
 } from '@angular/core';
 import { MatDialog, MatDialogRef } from '@angular/material/dialog';
 import { DetailedTermsComponent } from '@blockframes/contract/term/components/detailed/detailed.component';
-import { Movie } from '@blockframes/model';
+import { Holdback, Movie } from '@blockframes/model';
 import { FormTableComponent } from '@blockframes/ui/form/table/form-table.component';
 import { FormList } from '@blockframes/utils/form';
 import { Scope } from '@blockframes/utils/static-model';
-import { Holdback } from '../../+state/contract.model';
 import { HoldbackForm } from '../form';
 
 @Component({
