+++ libs/contract/src/lib/contract/holdback/pipes/colliding-holdback.pipe.ts
@@ -1,10 +1,7 @@
-
 import { Pipe, PipeTransform, NgModule } from '@angular/core';
-
 import { BucketTerm } from '@blockframes/contract/term/+state';
 import { getCollidingHoldbacks } from '@blockframes/contract/avails/avails';
-
-import { Holdback } from '../../+state';
+import { Holdback } from '@blockframes/model';
 
 @Pipe({ name: 'collidingHoldbacks' })
 export class CollidingHoldbacksPipe implements PipeTransform {
