+++ libs/contract/src/lib/contract/list/internal-sales/internal-sales.component.ts
@@ -1,5 +1,4 @@
 import { Component, ChangeDetectionStrategy, OnInit, Input } from '@angular/core';
-import { Contract, ContractStatus, Sale } from '@blockframes/contract/contract/+state';
 import { ActivatedRoute, Router } from '@angular/router';
 import { filter, map, startWith } from 'rxjs/operators';
 import { FormControl } from '@angular/forms';
@@ -7,6 +6,7 @@ import { BehaviorSubject, Observable } from 'rxjs';
 import { DynamicTitleService } from '@blockframes/utils/dynamic-title/dynamic-title.service';
 import { isInitial } from '@blockframes/contract/negotiation/utils';
 import { Negotiation } from '@blockframes/contract/negotiation/+state/negotiation.firestore';
+import { Contract, ContractStatus, Sale } from '@blockframes/model';
 
 function capitalize(text: string) {
   return `${text[0].toUpperCase()}${text.substring(1)}`;
