+++ libs/contract/src/lib/offer/list/list.component.ts
@@ -3,8 +3,7 @@ import { Offer } from '@blockframes/contract/offer/+state';
 import { Observable } from 'rxjs';
 import { FormControl } from '@angular/forms';
 import { startWith } from 'rxjs/operators';
-import { Contract } from '@blockframes/contract/contract/+state';
-import { Movie } from '@blockframes/model';
+import { Contract, Movie } from '@blockframes/model';
 import { DynamicTitleService } from '@blockframes/utils/dynamic-title/dynamic-title.service';
 import { ActivatedRoute, Router } from '@angular/router';
 import { Negotiation } from '@blockframes/contract/negotiation/+state/negotiation.firestore';
