+++ libs/contract/src/lib/term/pipes.ts
@@ -1,8 +1,8 @@
 import { NgModule, Pipe, PipeTransform } from "@angular/core";
+import { Contract } from "@blockframes/model";
 import { of } from "rxjs";
 import { BucketContract } from "@blockframes/model";
 import { TermService } from "../term/+state";
-import { Contract } from "../contract/+state";
 
 const isBucketContract = (contract: Contract | BucketContract): contract is BucketContract => {
   return 'terms' in contract;
