+++ libs/import/src/lib/utils.ts
@@ -1,6 +1,5 @@
 import { MovieService } from '@blockframes/movie/+state/movie.service';
-import { Movie, Organization, User } from '@blockframes/model';
-import { Mandate, Sale } from '@blockframes/contract/contract/+state/contract.model';
+import { Movie, Organization, User, Mandate, Sale } from '@blockframes/model';
 import { OrganizationService } from '@blockframes/organization/+state';
 import { Term } from '@blockframes/contract/term/+state/term.model';
 import { SheetTab, ValueWithError } from '@blockframes/utils/spreadsheet';
