+++ libs/import/src/lib/view-extracted-elements/contract/utils.ts
@@ -22,10 +22,8 @@ import { UserService } from '@blockframes/user/+state';
 import { OrganizationService } from '@blockframes/organization/+state';
 import { Term } from '@blockframes/contract/term/+state/term.firestore';
 import { Language, Media, Territory } from '@blockframes/utils/static-model';
-import { MovieLanguageSpecification, User } from '@blockframes/model';
-import { Mandate, Sale } from '@blockframes/contract/contract/+state/contract.firestore';
+import { createMandate, createSale, Mandate, MovieLanguageSpecification, Sale, User } from '@blockframes/model';
 import { ContractService } from '@blockframes/contract/contract/+state/contract.service';
-import { createMandate, createSale } from '@blockframes/contract/contract/+state/contract.model';
 import {
   extract, ExtractConfig, getStaticList, SheetTab, getGroupedList
 } from '@blockframes/utils/spreadsheet';
