+++ libs/model/src/index.ts
@@ -3,6 +3,7 @@ export * from './lib/user';
 export * from './lib/invitation';
 export * from './lib/notification';
 export * from './lib/organisation';
-export * from './lib/event';
 export * from './lib/permissions';
 export * from './lib/bucket';
+export * from './lib/contract';
+export * from './lib/event';
