+++ libs/model/src/lib/bucket.ts
@@ -1,12 +1,9 @@
-import { AvailsFilter, CalendarAvailsFilter, MapAvailsFilter } from '@blockframes/contract/avails/avails';
-import { Holdback, MailContract } from '@blockframes/contract/contract/+state/contract.firestore';
-import { BucketTerm, Term } from '@blockframes/contract/term/+state/term.firestore';
-import { createHoldback, Mandate } from '@blockframes/contract/contract/+state';
-import { createLanguageKey } from '@blockframes/model';
+import { BucketTerm } from '@blockframes/contract/term/+state/term.firestore';
+import { createLanguageKey, createHoldback, Holdback, MailContract } from '@blockframes/model';
 import type { MovieCurrency } from '@blockframes/utils/static-model';
 import type firebase from 'firebase';
 
-export interface Bucket<T extends Date | firebase.firestore.Timestamp = Date>  {
+export interface Bucket<T extends Date | firebase.firestore.Timestamp = Date> {
   id: string;
   currency: MovieCurrency;
   /** One contract per orgId / titleId / parent terms Id */
@@ -43,17 +40,6 @@ export interface MailBucket {
   uid?: string;
 }
 
-export function toBucketTerm(
-  avail: AvailsFilter | MapAvailsFilter | CalendarAvailsFilter
-): BucketTerm {
-  return createBucketTerm({
-    medias: avail.medias,
-    duration: 'duration' in avail ? avail.duration : undefined,
-    territories: 'territories' in avail ? avail.territories : undefined,
-    exclusive: avail.exclusive,
-  });
-}
-
 export function createBucketTerm(params: Partial<BucketTerm> = {}): BucketTerm {
   return {
     territories: [],
@@ -76,19 +62,6 @@ export function createBucketContract(params: Partial<BucketContract> = {}): Buck
     terms: params.terms?.map(createBucketTerm) ?? [],
     holdbacks: params.holdbacks?.map(createHoldback) ?? [],
   };
-}
-
-export function toBucketContract(
-  contract: Mandate,
-  term: Term<Date>,
-  avails: AvailsFilter | MapAvailsFilter | CalendarAvailsFilter
-): BucketContract {
-  return createBucketContract({
-    titleId: contract.titleId,
-    orgId: contract.sellerId,
-    parentTermId: term.id,
-    terms: [toBucketTerm(avails)],
-  });
 }
 
 export function createBucket(params: Partial<Bucket> = {}): Bucket {
