+++ libs/notification/src/lib/+state/notification.service.ts
@@ -10,7 +10,8 @@ import {
   isScreening,
   Notification,
   Organization,
-  OrganizationDocument
+  OrganizationDocument,
+  Contract
 } from '@blockframes/model';
 import { OrganizationService } from '@blockframes/organization/+state';
 import { toDate } from '@blockframes/utils/helpers';
@@ -25,8 +26,7 @@ import { UserService } from '@blockframes/user/+state';
 import { EventService } from '@blockframes/event/+state';
 import { ModuleGuard } from '@blockframes/utils/routes/module.guard';
 import { APP } from '@blockframes/utils/routes/utils';
-import { OfferService } from '@blockframes/contract/offer/+state';
-import { Contract, ContractService } from '@blockframes/contract/contract/+state';
+import { ContractService } from '@blockframes/contract/contract/+state';
 
 interface NotificationState extends EntityState<Notification>, ActiveState<string> {}
 @Injectable({ providedIn: 'root' })
